marketing-agent/server/pyproject.toml
[tool.poetry]
name = "crb-marketing-agent"
version = "0.1.0"
description = ""
authors = [
    "Leo Kim <leo.kim@cerebras.net>",
    "Kunal Patel <kunal@thatone.ai>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
cerebras-cloud-sdk = {url = "https://cerebras-cloud-sdk.s3.amazonaws.com/production/cerebras_cloud_sdk-0.7.0-py3-none-any.whl"}
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
openai = "^1.42.0"
fireworks-ai = "^0.15.0"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
groq = "^0.9.0"
together = "^1.2.7"

[tool.poetry.group.dev.dependencies]
aiohttp = "^3.10.5"
websockets = "^13.0"

[virtualenvs]
in-project = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
